requirements.txt
# ============================================================================
# Alexandria RAG System - Python Dependencies
# ============================================================================
#
# ⚠️  CRITICAL: This is the OFFICIAL Python dependency file (.txt NOT .md)
# ⚠️  DO NOT DELETE or rename to .md - pip install -r requirements.txt expects .txt
# ⚠️  Python tools (pip, PyCharm, Docker, CI/CD) all look for requirements.txt
#
# Python 3.14+ required

# Vector Database
qdrant-client>=1.7.1

# Embedding & Semantic Analysis
sentence-transformers>=2.3.1  # Multi-model: minilm (384-dim), bge-large (1024-dim)
torch>=2.0.0  # Required by sentence-transformers
# For CUDA GPU support, install: pip install torch --index-url https://download.pytorch.org/whl/cu121
huggingface_hub[hf_xet]  # Xet protocol for faster model downloads (chunked, resumable)
numpy>=1.24.0  # Universal chunking semantic analysis
scikit-learn>=1.3.0  # Cosine similarity for semantic chunking

# Book Parsing - EPUB
EbookLib==0.18

# Book Parsing - PDF
PyMuPDF>=1.24.0

# HTML Parsing (for EPUB content extraction)
beautifulsoup4==4.12.2
lxml>=4.9.3

# Optional: MOBI support (requires additional setup)
# mobi==0.3.3

# HTTP Client
requests>=2.31.0  # OpenRouter API calls

# Utilities
tqdm==4.66.1  # Progress bars (disabled globally via TQDM_DISABLE=1)

# MCP Server
mcp>=1.0.0  # Model Context Protocol SDK for Claude Code integration

# GUI
streamlit>=1.29.0  # Web interface (binds to 0.0.0.0:8501)

# Observability & Evaluation
arize-phoenix>=4.0.0  # RAG observability, tracing, evaluation dashboard
opentelemetry-api>=1.20.0  # OpenTelemetry tracing API
opentelemetry-sdk>=1.20.0  # OpenTelemetry SDK
opentelemetry-instrumentation>=0.41b0  # Auto-instrumentation base

# Development (Recommended - Not Yet Implemented)
pytest==7.4.3  # Testing framework
pytest-cov==4.1.0  # Coverage reports
pytest-playwright>=0.4.0  # Browser/UI testing with Playwright
black==23.12.1  # Code formatter
flake8==7.0.0  # Linter
